Use full path to reference userland and framework objects.
You now have to use the full path when referencing anything that is
automatically loaded into memory.

Example:

```javascript

//modules/admin/modules/users/.web/routes.js

export default {

 '/': {
       get: {
           action: 'admin.users.Main',
           middleware: ['admin.audit']
      }
}

};
```
A bit tedious, but it allows us to avoid name collisions in
the future. If using a module that we don't want to
modify at all (example; a bit submodule) then they will have
to be in the main `modules` folder to avoid having to change
all the paths.